pyproject.toml
[tool.poetry]
name = "signature_processing_api"
version = "0.1.0"
description = ""
authors = ["Matheus Parracho <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "processing" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
flask = "3.1.1"
torch = "2.2.1"
torchvision = "0.17.1"
Pillow = "11.0.0"
albumentations = "1.4.15"
albucore = "0.0.15"
flasgger = "0.9.7.1"
numpy = "1.26.4"
gunicorn = "21.2.0"

[tool.poetry.scripts]
signature-server = "app:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"